+++ source_gen/source_gen/source_gen/lib/src/type_checker.dart
@@ -302,7 +302,7 @@ class UnresolvedAnnotationException implements Exception {
   ) {
     try {
       final parsedLibrary = annotatedElement.session!
-              .getParsedLibraryByElement2(annotatedElement.library!)
+              .getParsedLibraryByElement(annotatedElement.library!)
           as ParsedLibraryResult;
       final declaration = parsedLibrary.getElementDeclaration(annotatedElement);
       if (declaration == null) {